Make math tasks support double instead of int

+++ lab-01/by/toharrius/quizer/tasks/math_tasks/AbstractMathTask.java
@@ -3,11 +3,11 @@ package by.toharrius.quizer.tasks.math_tasks;
 import java.util.EnumSet;
 
 public abstract class AbstractMathTask implements MathTask {
-    protected final int a;
+    protected final double a;
     protected final MathOperation op;
-    protected final int b;
-    protected final int c;
-    public AbstractMathTask(int a, MathOperation op, int b) {
+    protected final double b;
+    protected final double c;
+    public AbstractMathTask(double a, MathOperation op, double b) {
         this.a = a;
         this.op = op;
         this.b = b;
@@ -15,10 +15,10 @@ public abstract class AbstractMathTask implements MathTask {
     }
 
     protected abstract static class Generator implements MathTask.Generator {
-        private final int minNumber;
-        private final int maxNumber;
+        private final double minNumber;
+        private final double maxNumber;
         private final EnumSet<MathOperation> allowed;
-        public Generator(int minNumber, int maxNumber,
+        public Generator(double minNumber, double maxNumber,
                          EnumSet<MathOperation> allowed) {
             this.minNumber = minNumber;
             this.maxNumber = maxNumber;
@@ -26,12 +26,12 @@ public abstract class AbstractMathTask implements MathTask {
         }
 
         @Override
-        public int getMinNumber() {
+        public double getMinNumber() {
             return minNumber;
         }
 
         @Override
-        public int getMaxNumber() {
+        public double getMaxNumber() {
             return maxNumber;
         }
